Support saving with a WebDAV style PUT request
The motivation is to support FeatherWiki on Tiddlyhost, but in
theory this could work with any WebDAV server.

The techniques for saving and for detecting if the server supports
put saves are borrowed from TiddlyWiki's put saver, see here:
https://github.com/Jermolene/TiddlyWiki5/blob/39e4e69ae7/core/modules/savers/put.js

When the put save is enabled the change flag and the "save needed"
indicator are associated with the put save button only, since I
think that's more intuitive and it avoids potential confusion if the
change flag is cleared after a download save, but the server version
is still out of date.

Closes #4