Support round-trip saving with "bare bones" html
I'm calling it "bare bones", but what I mean is a small html file
that does not contain the main script or the main stylesheet.
Instead it references them externally.

With the change, the html generation can detect if the css or
javascript is being referenced externally, and output the
appropriate html to keep it the same.

(The reason I'm interested in this is because I think it could work
well on Tiddlyhost since the files would load faster and the amount
of data being "PUT" to the server on each save would be smaller.)